+++ src/Modules/OrchardCore.Commerce/Services/NumericProductAttributeProvider.cs
@@ -63,7 +63,7 @@ public class NumericProductAttributeProvider : IProductAttributeProvider
         var type = _contentDefinitionManager.GetTypeDefinition(productPart.ContentItem.ContentType);
         foreach (var attribute in numericAttributesList)
         {
-            var (attributePartDefinition, attributeFieldDefinition) = GetFieldDefinition(
+            var (attributePartDefinition, attributeFieldDefinition) = _productAttributeService.GetFieldDefinition(
                 type, type.Name + "." + attribute.Name);
 
             var defaultValue = ((attribute.Settings as NumericProductAttributeFieldSettings).DefaultValue ?? 0)
@@ -78,23 +78,5 @@ public class NumericProductAttributeProvider : IProductAttributeProvider
 
             attributesList.Add(matchingAttribute);
         }
-    }
-
-    private static (ContentTypePartDefinition PartDefinition, ContentPartFieldDefinition FieldDefinition)
-        GetFieldDefinition(ContentTypeDefinition type, string attributeName)
-    {
-        var partAndField = attributeName.Split('.');
-        var partName = partAndField[0];
-        var fieldName = partAndField[1];
-
-        return type
-            .Parts
-            .Where(partDefinition => partDefinition.Name == partName)
-            .SelectMany(partDefinition => partDefinition
-                .PartDefinition
-                .Fields
-                .Select(fieldDefinition => (PartDefinition: partDefinition, FieldDefinition: fieldDefinition))
-                .Where(pair => pair.FieldDefinition.Name == fieldName))
-            .FirstOrDefault();
     }
 }
